Added an F4MatrixBuilder and an F4Reducer - they do not work yet.
The new code in this commit compiles but if you specify the option to
use the F4Reducer (reducer #25) you get a segmentation fault for
classically reducing an S-pair. The other F4 reducer methods work only
because they currently are actually just letting another reducer do
the work. Next up: testing and fixing.